+++ include/core.h
@@ -2,4 +2,4 @@
 
 #include "net.h"
 #include "layers.h"
-#include "modules/generic_layers.h"
+#include "generic_layers.hpp"
