+++ src/modules/generic_layers.hpp
@@ -1,6 +1,5 @@
 #pragma once
 #include "layers.h"
-#include "rng_helpers.h"
 
 namespace rum
 {
